v0.16.3 fixes bug where null keys can cause bad merges
